requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.135
botocore==1.29.135
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
cron-descriptor==1.3.0
cycler==0.11.0
Django==4.2.1
django-celery-beat==2.5.0
django-cors-headers==4.0.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.39.4
freetype-py==2.3.0
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.4
kombu==5.2.4
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
paytm==0.1.8
pdfkit==1.0.0
pika==1.3.2
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycairo==1.23.0
pycrypto==2.6.1
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
reportlab==4.0.0
requests==2.30.0
rlPyCairo==0.2.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
